Massive logging improvements: 1. Grab the logs from the containers themselves (using docker) and integrate them into the logs for the test 2. Log to a file so it can be tailed by tools like `lnav`. (Highly recommend lnav!!!!) 3. Lots of wrappers and contexts around different methods in testermint 4. In inferenced itself, start using the loggers given by the CosmosSDK with levels. 5. One note: Since CosmosSDK positively DUMPS out info logs, I created an "INFO+" mechanism so we don't log everything that's coming as INFO (those get mapped to Debug) in the tests. 6. Also implemented most of the settle payments.

+++ testermint/src/main/kotlin/data/status.kt
@@ -33,11 +33,11 @@ data class Other(
 data class SyncInfo(
     val latestBlockHash: String,
     val latestAppHash: String,
-    val latestBlockHeight: Int,
+    val latestBlockHeight: Long,
     val latestBlockTime: String,
     val earliestBlockHash: String,
     val earliestAppHash: String,
-    val earliestBlockHeight: Int,
+    val earliestBlockHeight: Long,
     val earliestBlockTime: Instant,
     val catchingUp: Boolean
 )
